requirements.txt
httpx
google-adk